wpt_babel/content-security-policy/unsafe-hashes/script_event_handlers_denied_wrong_hash.html

<!DOCTYPE HTML>
<html>
 <head>
  <meta content="script-src 'unsafe-hashes' 'nonce-abc' 'sha256-thisdoesnotmatch'; img-src *;" http-equiv="Content-Security-Policy"/>
  <title>
   Event handlers should be not allowed if a matching hash is not present
  </title>
  <script nonce="abc" src="/resources/testharness.js">
  </script>
  <script nonce="abc" src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <div id="log">
  </div>
  <script nonce="abc">
   "use strict";

var t1 = async_test("Test that the inline event handler is not allowed to run");
window.addEventListener('securitypolicyviolation', t1.step_func_done(function (e) {
  assert_equals(e.violatedDirective, 'script-src-attr');
  assert_equals(e.blockedURI, 'inline');
}));
  </script>
  <img onload='t1.unreached_func("Should not have executed handler");' src="../support/pass.png"/>
 </body>
</html>
